norme libft/libprintf
