+++ libft/ft_atod.c
@@ -6,13 +6,13 @@
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/19 16:39:09 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/04/29 18:20:01 by ebaudet          ###   ########.fr       */
+/*   Updated: 2019/04/29 18:51:00 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_pow10(unsigned int y)
+static double	ft_pow10(unsigned int y)
 {
 	double result;
 
@@ -25,23 +25,10 @@ double	ft_pow10(unsigned int y)
 	return (result);
 }
 
-double	ft_atod(const char *str)
+static double	ft_atod_decimals(const char *str, double value, int sign)
 {
-	double			value;
-	int				sign;
 	unsigned int	i;
 
-	value = 0;
-	sign = (*str == '-') ? -1 : 1;
-	while (ft_isspace(*str))
-		str++;
-	if (*str == '-' || *str == '+')
-		str++;
-	while (ft_isdigit(*str))
-	{
-		value = value * 10 + (*str - '0');
-		str++;
-	}
 	if (*str == '.')
 	{
 		(str)++;
@@ -54,4 +41,23 @@ double	ft_atod(const char *str)
 	}
 	str += 2;
 	return (value * sign);
+}
+
+double			ft_atod(const char *str)
+{
+	double			value;
+	int				sign;
+
+	value = 0;
+	sign = (*str == '-') ? -1 : 1;
+	while (ft_isspace(*str))
+		str++;
+	if (*str == '-' || *str == '+')
+		str++;
+	while (ft_isdigit(*str))
+	{
+		value = value * 10 + (*str - '0');
+		str++;
+	}
+	return (ft_atod_decimals(str, value, sign));
 }
