+++ libft/ft_printf/pf_buffer.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   buffer.c                                           :+:      :+:    :+:   */
+/*   pf_buffer.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/11 14:21:13 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/02/11 22:23:08 by ebaudet          ###   ########.fr       */
+/*   Updated: 2019/04/29 18:54:29 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
