+++ libft/ft_printf/pf_handler.c
@@ -6,36 +6,38 @@
 /*   By: ebaudet <ebaudet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/28 17:55:04 by ebaudet           #+#    #+#             */
-/*   Updated: 2019/04/29 18:33:11 by ebaudet          ###   ########.fr       */
+/*   Updated: 2019/04/29 19:48:41 by ebaudet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 #include "libft.h"
 
+static const t_call_handler	g_call_handler[] = {
+	{.value = P_FLAGS, .handle = flag_handler},
+	{.value = P_WIDTH, .handle = width_handler},
+	{.value = P_PRECISION, .handle = precision_handler},
+	{.value = P_LENGTH, .handle = length_handler},
+	{.value = P_TYPE, .handle = type_handler},
+	{.value = P_MODULO, .handle = modulo_handler},
+	{.value = NULL, .handle = NULL}
+};
+
 int		call_handler(const char *format, t_ftprintf *t, t_params *params)
 {
-	static t_call_handler	handler[] = {
-		{.value = P_FLAGS, .handle = flag_handler},
-		{.value = P_WIDTH, .handle = width_handler},
-		{.value = P_PRECISION, .handle = precision_handler},
-		{.value = P_LENGTH, .handle = length_handler},
-		{.value = P_TYPE, .handle = type_handler},
-		{.value = P_MODULO, .handle = modulo_handler},
-		{.value = NULL, .handle = NULL}
-	};
 	int						i;
 	int						do_stuff;
 
 	do_stuff = 0;
 	i = -1;
-	while (handler[++i].value != NULL)
+	while (g_call_handler[++i].value != NULL)
 	{
-		if (format[t->i] && ft_strchr(handler[i].value, format[t->i]))
+		if (format[t->i] && ft_strchr(g_call_handler[i].value, format[t->i]))
 		{
-			(handler[i].handle)(&format[t->i], t, params);
+			(g_call_handler[i].handle)(&format[t->i], t, params);
 			do_stuff++;
-			if (i == 4 || i == 5)
+			if (ft_strcmp(g_call_handler[i].value, P_TYPE)
+				|| ft_strcmp(g_call_handler[i].value, P_MODULO))
 				return (-1);
 		}
 	}
@@ -115,65 +117,4 @@ void	precision_handler(const char *format, t_ftprintf *t, t_params *params)
 	ft_strncpy(number, &format[1], last_number - &format[1]);
 	(t->i) += last_number - format;
 	params->precision = (int)ft_atoi(number);
-}
-
-void	length_handler(const char *format, t_ftprintf *t, t_params *params)
-{
-	if (format[0] == 'h')
-	{
-		if (format[1] && format[1] == 'h' && t->i++)
-			params->length = params->length ? params->length : HH;
-		else
-			params->length = params->length ? params->length : H;
-	}
-	else if (format[0] == 'l')
-	{
-		if (format[1] && format[1] == 'l' && t->i++)
-			params->length = params->length ? params->length : LL;
-		else
-			params->length = params->length ? params->length : L;
-	}
-	else if (format[0] == 'L')
-		params->length = params->length ? params->length : LD;
-	else if (format[0] == 'z')
-		params->length = params->length ? params->length : Z;
-	else if (format[0] == 'j')
-		params->length = params->length ? params->length : J;
-	else if (format[0] == 't')
-		params->length = params->length ? params->length : T;
-	else
-		params->length = params->length ? params->length : NONE_LENGH;
-	t->i++;
-}
-
-static const t_handler	g_handler[] = {
-	{.value = 'c', .handle = type_c},
-	{.value = 's', .handle = type_s},
-	{.value = 'p', .handle = type_p},
-	{.value = 'd', .handle = type_d},
-	{.value = 'i', .handle = type_d},
-	{.value = 'o', .handle = type_o},
-	{.value = 'u', .handle = type_u},
-	{.value = 'x', .handle = type_x},
-	{.value = 'X', .handle = type_x_cap},
-	{.value = 'f', .handle = type_f},
-	{.value = 'k', .handle = type_k},
-	{.value = 'b', .handle = type_b},
-	{.value = 0, .handle = NULL}
-};
-
-void	type_handler(const char *format, t_ftprintf *t, t_params *params)
-{
-	int		i;
-
-	i = -1;
-	while (g_handler[++i].value != 0)
-	{
-		if (format[0] == g_handler[i].value)
-		{
-			(*g_handler[i].handle)(t, params);
-			t->i++;
-			return ;
-		}
-	}
 }
